refactor: separate coverage input validation
